car_color_classification/requirements.txt
gdown==4.7.1
natsort==8.4.0
numpy==1.23.4
pandas==1.5.1
Pillow==10.0.1
PyYAML==6.0.1
scikit_learn==1.3.1
tqdm==4.64.1
